+++ src/main/java/gift/repository/ProductH2Repository.java
@@ -1,0 +1,2 @@
+package gift.repository;public class ProductH2Repository {
+}
