+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,38 +1,2 @@
-package gift.repository;
-
-import gift.domain.Product;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class ProductRepository {
-
-    private final Map<Long, Product> products = new HashMap<>();
-    private Long Id = 1L;
-
-    public Optional<Product> findById(Long id) {
-        return Optional.ofNullable(products.get(id));
-    }
-
-    public Iterable<Product> findAll() {
-        return products.values();
-    }
-
-    public Product save(Product product) {
-        if (product.getId() == null) {
-            product.setId(Id++);
-        }
-        products.put(product.getId(), product);
-        return product;
-    }
-
-    public void deleteById(Long id) {
-        Id--;
-        products.remove(id);
-
-    }
+package gift.repository;public class ProductRepository {
 }
-
-
