+++ src/main/resources/templates/index.html
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
-<html lang="en">
+<html xmlns:th="http://www.thymeleaf.org">
 <head>
-  <meta charset="UTF-8">
-  <title>$Title$</title>
+  <title>Product Management</title>
+  <link rel="stylesheet" type="text/css" href="/style.css">
 </head>
 <body>
-$END$
+<h1>Welcome to Product Management System</h1>
+<a href="/products">Manage Products</a>
 </body>
 </html>
